moved core/parallel.py to parallel_utils.py
