+++ docs/examples/simulator_initialization.md
@@ -132,7 +132,7 @@ print("available CPUs YAQS would use:", available_cpus())
 print("default max_workers:          ", Simulator().max_workers)
 ```
 
-{func}`~mqt.yaqs.core.parallel.available_cpus` (re-exported as {func}`~mqt.yaqs.simulator.available_cpus`) is deliberately cgroup- and scheduler-aware. In priority order it honours:
+{func}`~mqt.yaqs.parallel_utils.available_cpus` (re-exported as {func}`~mqt.yaqs.simulator.available_cpus`) is deliberately cgroup- and scheduler-aware. In priority order it honours:
 
 1. `YAQS_MAX_WORKERS` (explicit user override; positive integer).
 2. `PYTEST_XDIST_WORKER` (returns `1` to avoid nested parallelism in tests).
@@ -175,7 +175,7 @@ The bar is suppressed regardless of `parallel`, so the same flag also silences s
 | `"spawn"` | Fresh interpreter per worker. Required on Windows/macOS; slower startup but more isolated.                      |
 
 ```{code-cell} ipython3
-from mqt.yaqs.core.parallel import get_parallel_context
+from mqt.yaqs.parallel_utils import get_parallel_context
 
 for choice in ("auto", "fork", "spawn"):
     try:
